chore: update queryLog trigger
